Organizing protocol extension types
